chatgpt_backend/requirements.txt
Django==4.1.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
openai==0.27.0
django-cors-headers==3.14.0